requirements.txt
streamlit~=1.17.0
pandas~=1.5.3
numpy~=1.24.2
pyyaml~=6.0
mysql-connector-python~=8.0.32